NXP-17008: cleanup

+++ nuxeo-features/nuxeo-platform-audit/nuxeo-platform-audit-core/src/test/java/org/nuxeo/ecm/platform/audit/TestDocumentAuditPageProvider.java
@@ -156,8 +156,8 @@ public class TestDocumentAuditPageProvider {
         int proxyEntriesCount = 19 - 5 + 1 + 1;
         assertEquals(proxyEntriesCount, entries.size());
 
-        assertEquals(Long.valueOf(startId).longValue(), entries.get(proxyEntriesCount - 1).getId());
-        assertEquals(Long.valueOf(startId + proxyEntriesCount + 1).longValue(), entries.get(0).getId());
+        assertEquals(startId, entries.get(proxyEntriesCount - 1).getId());
+        assertEquals(startId + proxyEntriesCount + 1, entries.get(0).getId());
 
         // Get version 1 history
         pp = pps.getPageProvider("DOCUMENT_HISTORY_PROVIDER", null, Long.valueOf(20), Long.valueOf(0),
@@ -168,8 +168,8 @@ public class TestDocumentAuditPageProvider {
         // creation + 5 updates + update + checkin + created
         int version1EntriesCount = 1 + 5 + 1 + 1 + 1;
         if (version1EntriesCount == entries.size()) {
-            assertEquals(Long.valueOf(startId).longValue(), entries.get(version1EntriesCount - 1).getId());
-            assertEquals(Long.valueOf(startId + version1EntriesCount - 1).longValue(), entries.get(0).getId());
+            assertEquals(startId, entries.get(version1EntriesCount - 1).getId());
+            assertEquals(startId + version1EntriesCount - 1, entries.get(0).getId());
         } else {
             // because update even may be 1ms behind checkin/created !
             assertEquals(version1EntriesCount - 1, entries.size());
@@ -185,8 +185,8 @@ public class TestDocumentAuditPageProvider {
         // creation + 5x2 updates + checkin/update + checkin + created
         int versin2EntriesCount = 1 + 5 * 2 + 1 + 1 + 1 + 1;
         assertEquals(versin2EntriesCount, entries.size());
-        assertEquals(Long.valueOf(startId).longValue(), entries.get(versin2EntriesCount - 1).getId());
-        assertEquals(Long.valueOf(startId + versin2EntriesCount).longValue(), entries.get(0).getId());
+        assertEquals(startId, entries.get(versin2EntriesCount - 1).getId());
+        assertEquals(startId + versin2EntriesCount, entries.get(0).getId());
 
     }
